Add a documentation about what can be done with the search input
