+++ rudder-core/src/main/scala/com/normation/rudder/services/quicksearch/QuickSearchDomain.scala
@@ -65,23 +65,6 @@ object QSBackend {
   final val all: Set[QSBackend] = sealerate.values[QSBackend]
 }
 
-
-/*
- * Objects on which we are able to perform
- * quicksearch
- */
-sealed trait QSObject { def name: String }
-
-object QSObject {
-  final case object Node      extends QSObject { override val name = "node" }
-  final case object Group     extends QSObject { override val name = "group" }
-  final case object Directive extends QSObject { override val name = "directive" }
-  final case object Parameter extends QSObject { override val name = "parameter" }
-  final case object Rule      extends QSObject { override val name = "rule" }
-
-  final val all: Set[QSObject] = sealerate.values[QSObject]
-}
-
 /*
  * List of attribute for each type of objects that can be query.
  * The name are the same that the corresponding name from
@@ -143,6 +126,128 @@ object QSAttribute {
 
 }
 
+
+
+/*
+ * Objects on which we are able to perform
+ * quicksearch
+ */
+sealed trait QSObject { def name: String; def attributes: Set[QSAttribute] }
+
+object QSObject {
+  import QSAttribute._
+  private[this] val commons: Set[QSAttribute] = Set(Name, Description, ShortDescription, LongDescription, IsEnabled)
+
+  final case object Node      extends QSObject { override val name = "node"
+                                                 override val attributes = commons ++ Set(NodeId, Fqdn, OsType, OsName
+                                                   , OsVersion, OsFullName, OsKernelVersion, OsServicePack, Arch, Ram
+                                                   , IpAddresses, PolicyServerId, Properties, RudderRoles)
+  }
+  final case object Group     extends QSObject { override val name = "group"
+                                                 override val attributes = commons ++ Set(GroupId, IsDynamic)
+  }
+  final case object Directive extends QSObject { override val name = "directive"
+                                                 override val attributes = commons ++ Set(DirectiveId, DirectiveVarName
+                                                   , DirectiveVarValue, TechniqueName, TechniqueVersion)
+  }
+  final case object Parameter extends QSObject { override val name = "parameter"
+                                                 override val attributes = commons ++ Set(ParameterName, ParameterValue)
+  }
+  final case object Rule      extends QSObject { override val name = "rule"
+                                                 override val attributes = commons ++ Set(RuleId, DirectiveIds, Targets)
+  }
+
+  final val all: Set[QSObject] = sealerate.values[QSObject]
+
+
+  // default sort for QuickSearchResult:
+  // - by type
+  // - then by name
+  def sortQSObject(a: QSObject, b:QSObject): Boolean = {
+
+    implicit class QSObjectOrder(o: QSObject) {
+      def order() = o match {
+        case Node      => 1
+        case Group     => 2
+        case Directive => 3
+        case Parameter => 4
+        case Rule      => 5
+      }
+    }
+    a.order <= b.order
+  }
+}
+
+/**
+ * Mapping between human and domain name.
+ * Mostly used in the query parser, but also in the
+ * documentation.
+ */
+final object QSMapping {
+  /**
+   * Mapping between a string and actual objects.
+   * We try to be kind with users: not case sensitive, not plural sensitive
+   */
+  val objectNameMapping = {
+    QSObject.all.map { obj =>
+      val n = obj.name.toLowerCase
+      (n -> obj) :: ( n + "s" -> obj) :: Nil
+    }.flatten.toMap
+  }
+
+  //set of names by attribute
+  val attributeNames = {
+    import QSAttribute._
+    val descriptions = Set(Description, LongDescription, ShortDescription).map( _.name ) ++ Set("descriptions")
+
+    QSAttribute.all.map { a => a match {
+      case Name              => (a, Set(Name.name, "name") )
+      case Description       => (a, descriptions)
+      case ShortDescription  => (a, descriptions)
+      case LongDescription   => (a, descriptions)
+      case IsEnabled         => (a, Set(IsEnabled.name, "enabled") )
+      case NodeId            => (a, Set(NodeId.name, "nodeid") )
+      case Fqdn              => (a, Set(Fqdn.name, "fqdn") )
+      case OsType            => (a, Set(OsType.name, "ostype", "os"))
+      case OsName            => (a, Set(OsName.name, "osname", "os") ) //not also full name because osFullname contains osName, not the reverse
+      case OsVersion         => (a, Set(OsVersion.name, "osversion", "os") )
+      case OsFullName        => (a, Set(OsFullName.name, "osfullname", OsName.name, "osname", "os") )
+      case OsKernelVersion   => (a, Set(OsKernelVersion.name, "oskernelversion", "oskernel", "kernel", "version", "os") )
+      case OsServicePack     => (a, Set(OsServicePack.name, "osservicepack", "ossp", "sp", "servicepack", "os") )
+      case Arch              => (a, Set(Arch.name, "architecture", "arch") )
+      case Ram               => (a, Set(Ram.name, "memory") )
+      case IpAddresses       => (a, Set(IpAddresses.name, "ip", "ips", "networkips") )
+      case PolicyServerId    => (a, Set(PolicyServerId.name, "policyserver") )
+      case Properties        => (a, Set(Properties.name, "node.props", "nodeprops", "node.properties", "nodeproperties") )
+      case RudderRoles       => (a, Set(RudderRoles.name, "serverrole", "serverroles", "role", "roles") )
+      case GroupId           => (a, Set(GroupId.name, "groupid") )
+      case IsDynamic         => (a, Set(IsDynamic.name, "isdynamic") )
+      case DirectiveId       => (a, Set(DirectiveId.name, "directiveid") )
+      case DirectiveVarName  => (a, Set(DirectiveVarName.name, "directivevar", "parameter", "parameters", "variable", "variables") )
+      case DirectiveVarValue => (a, Set(DirectiveVarValue.name,"directivevalue", "parameter", "parameters", "variable", "variables") )
+      case TechniqueName     => (a, Set(TechniqueName.name, "technique", "techniqueid") )
+      case TechniqueVersion  => (a, Set(TechniqueVersion.name, "technique", "techniqueid", "version") )
+      case ParameterName     => (a, Set(ParameterName.name, "parametername", "paramname", "name", "parameter", "param") )
+      case ParameterValue    => (a, Set(ParameterValue.name, "parametervalue", "paramvalue", "value", "parameter", "param") )
+      case RuleId            => (a, Set(RuleId.name, "ruleid") )
+      case DirectiveIds      => (a, Set(DirectiveIds.name, "directiveids", "id", "ids") )
+      case Targets           => (a, Set(Targets.name, "target", "group", "groups") )
+    } }
+  }.toMap
+
+  /*
+   * For attributes, we want to be a little more lenient than for objects, and have:
+   * - several names mapping to the same attribute. Ex: both (id, nodeid) map to NodeId.
+   * - a name mapping to several attributes. Ex. description map to (Description, LongDescription, ShortDescription)
+   * - for all name, also have the plural
+   */
+  val attributeNameMapping: Map[String, Set[QSAttribute]] = {
+    //given that mapping, build the map of name -> Set(attribute)
+    val byNames: Map[String, Map[String, QSAttribute]] = attributeNames.flatMap { case(a, names) => names.map( n => (n.toLowerCase,a) ) }.groupBy(_._1)
+    byNames.mapValues( _.map(_._2).toSet )
+  }
+
+}
 
 /**
  * And the domain for results: ids and full results.
