+++ rudder-core/src/main/scala/com/normation/rudder/services/quicksearch/QuickSearchQueryParser.scala
@@ -99,7 +99,10 @@ object QSRegexQueryParser extends RegexParsers {
         val missings = names -- oKeys -- aKeys
 
         if(missings.nonEmpty) {
-          val allNames = (objectNameMapping.keys.map( _.capitalize) ++ attributeNameMapping.keys).toSeq.sorted.mkString("', '")
+          val allNames = (
+               QSMapping.objectNameMapping.keys.map( _.capitalize)
+            ++ QSMapping.attributeNameMapping.keys
+          ).toSeq.sorted.mkString("', '")
           FailedBox(s"Query containts 'in' keys that are not knows: '${missings.mkString("','")}'. Please choose among '${allNames}'")
         } else {
           /*
@@ -221,66 +224,6 @@ object QSRegexQueryParser extends RegexParsers {
       }
   }
 
-  /**
-   * Mapping between a string and actual objects.
-   * We try to be kind with users: not case sensitive, not plural sensitive
-   */
-  private[this] val objectNameMapping = {
-    QSObject.all.map { obj =>
-      val n = obj.name.toLowerCase
-      (n -> obj) :: ( n + "s" -> obj) :: Nil
-    }.flatten.toMap
-  }
-
-  /*
-   * For attributes, we want to be a little more lenient than for objects, and have:
-   * - several names mapping to the same attribute. Ex: both (id, nodeid) map to NodeId.
-   * - a name mapping to several attributes. Ex. description map to (Description, LongDescription, ShortDescription)
-   * - for all name, also have the plural
-   */
-  private[this] val attributeNameMapping = {
-    import QSAttribute._
-    //set of names by attribute
-    val descriptions = Set(Description, LongDescription, ShortDescription).map( _.name ) ++ Set("descriptions")
-    val attributeNames = QSAttribute.all.map { a => a match {
-      case Name              => (a, Set(Name.name, "name") )
-      case Description       => (a, descriptions)
-      case ShortDescription  => (a, descriptions)
-      case LongDescription   => (a, descriptions)
-      case IsEnabled         => (a, Set(IsEnabled.name, "enabled") )
-      case NodeId            => (a, Set(NodeId.name, "nodeid") )
-      case Fqdn              => (a, Set(Fqdn.name, "fqdn") )
-      case OsType            => (a, Set(OsType.name, "ostype", "os"))
-      case OsName            => (a, Set(OsName.name, "osname", "os") ) //not also full name because osFullname contains osName, not the reverse
-      case OsVersion         => (a, Set(OsVersion.name, "osversion", "os") )
-      case OsFullName        => (a, Set(OsFullName.name, "osfullname", OsName.name, "osname", "os") )
-      case OsKernelVersion   => (a, Set(OsKernelVersion.name, "oskernelversion", "oskernel", "kernel", "version", "os") )
-      case OsServicePack     => (a, Set(OsServicePack.name, "osservicepack", "ossp", "sp", "servicepack", "os") )
-      case Arch              => (a, Set(Arch.name, "architecture", "arch") )
-      case Ram               => (a, Set(Ram.name, "memory") )
-      case IpAddresses       => (a, Set(IpAddresses.name, "ip", "ips", "networkips") )
-      case PolicyServerId    => (a, Set(PolicyServerId.name, "policyserver") )
-      case Properties        => (a, Set(Properties.name, "node.props", "nodeprops", "node.properties", "nodeproperties") )
-      case RudderRoles       => (a, Set(RudderRoles.name, "serverrole", "serverroles", "role", "roles") )
-      case GroupId           => (a, Set(GroupId.name, "groupid") )
-      case IsDynamic         => (a, Set(IsDynamic.name, "isdynamic") )
-      case DirectiveId       => (a, Set(DirectiveId.name, "directiveid") )
-      case DirectiveVarName  => (a, Set(DirectiveVarName.name, "directivevar", "parameter", "parameters", "variable", "variables") )
-      case DirectiveVarValue => (a, Set(DirectiveVarValue.name,"directivevalue", "parameter", "parameters", "variable", "variables") )
-      case TechniqueName     => (a, Set(TechniqueName.name, "technique", "techniqueid") )
-      case TechniqueVersion  => (a, Set(TechniqueVersion.name, "technique", "techniqueid", "version") )
-      case ParameterName     => (a, Set(ParameterName.name, "parametername", "paramname", "name", "parameter", "param") )
-      case ParameterValue    => (a, Set(ParameterValue.name, "parametervalue", "paramvalue", "value", "parameter", "param") )
-      case RuleId            => (a, Set(RuleId.name, "ruleid") )
-      case DirectiveIds      => (a, Set(DirectiveIds.name, "directiveids", "id", "ids") )
-      case Targets           => (a, Set(Targets.name, "target", "group", "groups") )
-    } }
-
-    //given that mapping, build the map of name -> Set(attribute)
-    val byNames: Map[String, Set[(String, QSAttribute)]] = attributeNames.flatMap { case(a, names) => names.map( n => (n.toLowerCase,a) ) }.groupBy(_._1)
-    byNames.mapValues( _.map(_._2))
-  }
-
   private[this] def getMapping[T](names: Set[String], map: Map[String, T]): (Set[T], Set[String]) = {
     val pairs = names.flatMap { name =>
       map.get(name.toLowerCase) match {
@@ -301,6 +244,7 @@ object QSRegexQueryParser extends RegexParsers {
    * inputs
    */
   private[this] def getObjects(names: Set[String]): (Set[QSObject], Set[String]) = {
+    import QSMapping._
     getMapping(names, objectNameMapping)
   }
 
@@ -309,6 +253,7 @@ object QSRegexQueryParser extends RegexParsers {
    * We try to be kind with users: not case sensitive, not plural sensitive
    */
   private[this] def getAttributes(names: Set[String]): (Set[QSAttribute], Set[String]) = {
+    import QSMapping._
     val (attrs, keys) = getMapping(names, attributeNameMapping)
     (attrs.flatten, keys)
   }
